Conversion of all ui controls to use similar format to RPi respository is underway
