+++ src/blue_and_pink_synth_editor/ui_controls/load_dialog.kv
@@ -1,0 +1,24 @@
+#:kivy 2.1.0
+
+<LoadDialog>:
+    orientation: 'vertical'
+    FileChooserListView:
+        id: filechooser
+        path: str(app._presets_directory_path)
+    BoxLayout:
+        size_hint_y: 1/8
+        Button:
+            text_size: self.size
+            font_size: self.height * 0.3
+            halign: 'center'
+            valign: 'middle'
+            text: "CANCEL"
+            on_release: root.cancel()
+
+        Button:
+            text_size: self.size
+            font_size: self.height * 0.3
+            halign: 'center'
+            valign: 'middle'
+            text: "OK"
+            on_release: root.load(filechooser.path, filechooser.selection)
