+++ src/blue_and_pink_synth_editor/ui_controls/save_dialog.kv
@@ -1,0 +1,39 @@
+#:kivy 2.1.0
+#:import C kivy.utils.get_color_from_hex
+
+<SaveDialog>:
+    text_input: text_input
+    orientation: "vertical"
+
+    TextInput:
+        id: text_input
+        text: root.default_filename
+        text_size: self.size
+        font_size: self.height * 0.3
+        halign: 'left'
+        valign: 'middle'
+        size_hint_y: 1/8
+        multiline: False
+
+    BoxLayout:
+        size_hint_y: 1/8
+        Button:
+            text_size: self.size
+            font_size: self.height * 0.3
+            halign: 'center'
+            valign: 'middle'
+            text: "CANCEL"
+            on_release: root.cancel()
+
+        Button:
+            text_size: self.size
+            font_size: self.height * 0.3
+            halign: 'center'
+            valign: 'middle'
+            text: "OK"
+            on_release: root.save(filechooser.path, filechooser.path + '/' + text_input.text)
+
+    FileChooserListView:
+        id: filechooser
+        path: str(app._presets_directory_path)
+        on_selection: text_input.text = self.selection and self.selection[0] or ''
