+++ src/blue_and_pink_synth_editor/ui_controls/value_control.py
